requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bandit==1.7.8
bce-python-sdk==0.9.5
bcrypt==4.2.0
beautifulsoup4==4.11.1
bleach==6.1.0
blinker==1.7.0
bs4==0.0.1
build==1.2.1
CacheControl==0.14.0
capstone==5.0.3
certifi==2022.6.15
cffi==1.16.0
chardet==5.0.0
charset-normalizer==2.0.12
cleo==2.1.0
click==8.1.7
colorama==0.4.4
colored-traceback==0.4.2
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.0
crashtest==0.4.1
cryptography==42.0.5
cycler==0.12.1
datasets==2.18.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.8
dnspython==2.2.1
docopt==0.6.2
docx2pdf==0.1.8
docx2txt==0.8
ds-store==1.3.1
dulwich==0.21.7
et-xmlfile==1.1.0
executing==2.1.0
exrex==0.10.5
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.13.1
fire==0.4.0
Flask==3.0.2
flask-babel==4.0.0
fonttools==4.50.0
frida==16.2.1
frida-tools==12.3.0
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.2
gevent==24.2.1
gitdb==4.0.11
GitHacker==1.1.7
GitPython==3.1.43
gmpy2==2.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.3
importlib_metadata==7.1.0
inexactsearch==1.0.2
installer==0.7.0
intervaltree==3.1.0
ipython==8.12.3
itsdangerous==2.1.2
jaraco.classes==3.4.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwt==1.3.1
keyring==24.3.1
kiwisolver==1.4.5
libnum==1.7.1
loguru==0.6.0
lxml==4.9.4
mac-alias==2.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.12.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
nltk==3.9.1
node-vm2==0.4.7
ntlm-auth==1.5.0
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-nvrtc-cu12==12.5.40
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==9.0.0.312
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nvjitlink-cu12==12.5.40
onnx==1.15.0
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddle2onnx==0.8.1
paddlefsl==1.1.0
paddlenlp==2.6.1
paddlepaddle==2.6.0
paddlepaddle-gpu==3.0.0b1
pandas==2.2.1
pandocfilters==1.5.1
paramiko==3.4.1
parso==0.8.4
pbr==6.0.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
pkginfo==1.10.0
platformdirs==4.2.0
plumbum==1.8.3
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prompt-toolkit==3.0.43
protobuf==3.20.2
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pwn==1.0
pwntools==4.13.1
pyarrow==15.0.1
pyarrow-hotfix==0.6
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pyelftools==0.31
Pygments==2.17.2
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyproject_hooks==1.0.0
pyreadline3==3.5.4
pyserial==3.5
PySocks==1.7.1
pyspellchecker==0.8.1
pyspnego==0.11.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.2.0
rapidfuzz==3.8.1
rarfile==4.1
referencing==0.35.1
regex==2024.9.11
requests==2.28.1
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
rich==13.7.1
ROPGadget==7.4
rpds-py==0.20.0
rpyc==6.0.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==3.0.2
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
silpa_common==0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.3.2
spark-parser==1.8.9
spellchecker==0.4
SQLAlchemy==1.3.22
sspilib==0.1.0
stack-data==0.6.3
starlette==0.36.3
stevedore==5.2.0
sympy==1.13.3
ten==0.1.4
tenacity==8.0.1
termcolor==1.1.0
threadpoolctl==3.3.0
tinycss2==1.3.0
tomlkit==0.12.4
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tornado==6.4.1
tqdm==4.64.0
traitlets==5.14.3
treelib==1.6.1
trove-classifiers==2024.4.10
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
uncompyle==2.0.0
uncompyle6==3.9.2
unicorn==2.0.1.post1
unix-ar==0.2.1
urllib3==1.26.9
uvicorn==0.28.0
virtualenv==20.25.3
visualdl==2.5.3
wavefile==1.6.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
win32-setctime==1.1.0
wxPython==4.2.1
xdis==6.1.1
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
zipp==3.18.1
zope.event==5.0
zope.interface==6.4
zstandard==0.23.0
